+++ RELEASE-NOTES.md
@@ -6,7 +6,7 @@
 * #213 StrExp.split()
 * #214 Parallel Aggregation of columns
 * #219 IndexedSeries.size(): NPE due to a race condtion
-* #221 Index.positions()
+* #221 Index.positions(), .positionsExcept()
 * #222 Refactor "Index.forLabel(..)" to "Index.of(..)" 
 
 ## Release 0.18